fix: fixed carpool map not rendering targets on first render
